need polishing

+++ libft/ft_memset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: oltolmac <oltolmac@student.42.fr>          +#+  +:+       +#+        */
+/*   By: olena <olena@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/02 11:54:04 by oltolmac          #+#    #+#             */
-/*   Updated: 2024/09/02 11:58:06 by oltolmac         ###   ########.fr       */
+/*   Updated: 2024/09/04 23:21:32 by olena            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *str, int c, size_t n)
 {
 	unsigned char *set;
-	int	i;
+	size_t	i;
 
 	i = 0;
 	set = (unsigned char *)str;
